지원자 관련 api 틀

+++ admin/src/main/java/kr/hs/entrydsm/admin/usecase/dto/ApplicantsInformationResponse.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Builder
 public class ApplicantsInformationResponse {
 
-    private Integer receiptCode;
+    private Long receiptCode;
 
     private String name;
